fix(tests): move firestore override and rename
