+++ test/mocks/providers/provider_create_account.dart
@@ -3,9 +3,9 @@ import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/views/navigation/routes.dart";
 import "package:proxima/views/pages/create_account_page.dart";
 
-import "../../services/firestore/testing_firestore_provider.dart";
-import "../overrides/mock_auth_providers.dart";
-import "../overrides/mock_home_view_model.dart";
+import "../overrides/override_auth_providers.dart";
+import "../overrides/override_firestore.dart";
+import "../overrides/override_home_view_model.dart";
 
 final ProviderScope createAccountPageProvider = ProviderScope(
   overrides: firebaseMocksOverrides +
