+++ test/mocks/providers/provider_new_post_page.dart
@@ -5,8 +5,8 @@ import "package:proxima/services/geolocation_service.dart";
 import "package:proxima/viewmodels/login_view_model.dart";
 import "package:proxima/views/pages/new_post/new_post_page.dart";
 
-import "../../services/firestore/testing_firestore_provider.dart";
 import "../data/mock_firestore_user.dart";
+import "../overrides/override_firestore.dart";
 import "../services/mock_geo_location_service.dart";
 import "../services/mock_post_repository_service.dart";
